+++ frontend/src/components/graph/hooks/node/useHandlePositions.ts
