+++ frontend/src/components/graph/hooks/node/useIsZoomedOut.ts
