+++ frontend/src/components/graph/nodes/ImageNode.tsx
@@ -5,8 +5,8 @@ import type { NodeProps, Node } from "@xyflow/react";
 import { ImageIcon, Film, X } from "lucide-react";
 import type { FlowNodeData } from "../../../lib/graphUtils";
 import { buildHandleId } from "../../../lib/graphUtils";
-import { useNodeData } from "../hooks/useNodeData";
-import { useNodeCollapse } from "../hooks/useNodeCollapse";
+import { useNodeData } from "../hooks/node/useNodeData";
+import { useNodeCollapse } from "../hooks/node/useNodeCollapse";
 import { getAssetUrl } from "../../../lib/api";
 import { MediaPicker, isVideoAsset } from "../../MediaPicker";
 import { NodeCard, NodeHeader, NODE_TOKENS, collapsedHandleStyle } from "../ui";
