+++ frontend/src/components/graph/utils/connectionValidation.ts
@@ -1,0 +1,328 @@
+/**
+ * Pure-function connection-validation logic.
+ *
+ * Extracted from useConnectionLogic so it can be unit-tested independently
+ * and keeps the hook thin.
+ */
+
+import type { Connection, Edge, Node } from "@xyflow/react";
+import { parseHandleId } from "../../../lib/graphUtils";
+import type { FlowNodeData } from "../../../lib/graphUtils";
+import { resolveSourceType } from "../hooks/connection/typeResolution";
+import { BOUNDARY_INPUT_ID, BOUNDARY_OUTPUT_ID } from "../hooks/subgraph/useGraphNavigation";
+
+// ---------------------------------------------------------------------------
+// Target-node rules (param → param).  Each entry maps a target nodeType (or
+// a special key) to a predicate that decides whether a given sourceType is
+// acceptable for that target handle.
+// ---------------------------------------------------------------------------
+
+type TargetRule = (ctx: {
+  sourceType: string;
+  targetParsedName: string;
+  targetNode: Node<FlowNodeData>;
+}) => boolean | undefined; // undefined = "this rule doesn't apply, fall through"
+
+/**
+ * Ordered list of rules evaluated when both handles are "param".
+ * The first rule that returns a boolean wins.  `undefined` means "skip".
+ */
+const TARGET_RULES: TargetRule[] = [
+  // __vace param – only vace type
+  ({ sourceType, targetParsedName }) => {
+    if (targetParsedName === "__vace") return sourceType === "vace";
+    return undefined;
+  },
+  // vace source can only go to __vace
+  ({ sourceType, targetParsedName }) => {
+    if (sourceType === "vace") return targetParsedName === "__vace";
+    return undefined;
+  },
+  // vace node targets
+  ({ sourceType, targetParsedName, targetNode }) => {
+    if (targetNode.data.nodeType !== "vace") return undefined;
+    if (targetParsedName === "video") {
+      if (sourceType !== "video_path") return false;
+      return !(
+        targetNode.data.vaceRefImage ||
+        targetNode.data.vaceFirstFrame ||
+        targetNode.data.vaceLastFrame
+      );
+    }
+    if (["ref_image", "first_frame", "last_frame"].includes(targetParsedName)) {
+      if (sourceType !== "string") return false;
+      return !targetNode.data.vaceVideo;
+    }
+    return false;
+  },
+  // __prompt param – only string
+  ({ sourceType, targetParsedName }) => {
+    if (targetParsedName === "__prompt") return sourceType === "string";
+    return undefined;
+  },
+  // math node – number for a / b
+  ({ sourceType, targetParsedName, targetNode }) => {
+    if (targetNode.data.nodeType !== "math") return undefined;
+    return (
+      sourceType === "number" &&
+      (targetParsedName === "a" || targetParsedName === "b")
+    );
+  },
+  // bool node – number for input
+  ({ sourceType, targetParsedName, targetNode }) => {
+    if (targetNode.data.nodeType !== "bool") return undefined;
+    return sourceType === "number" && targetParsedName === "input";
+  },
+  // string-control in switch mode
+  ({ sourceType, targetParsedName, targetNode }) => {
+    if (
+      targetNode.data.nodeType !== "control" ||
+      targetNode.data.controlType !== "string" ||
+      targetNode.data.controlMode !== "switch"
+    )
+      return undefined;
+    if (targetParsedName.startsWith("item_")) return sourceType === "number";
+    if (targetParsedName.startsWith("str_")) return sourceType === "string";
+    return false;
+  },
+  // slider / knobs / xypad – number only
+  ({ sourceType, targetNode }) => {
+    if (
+      targetNode.data.nodeType === "slider" ||
+      targetNode.data.nodeType === "knobs" ||
+      targetNode.data.nodeType === "xypad"
+    )
+      return sourceType === "number";
+    return undefined;
+  },
+  // tuple
+  ({ sourceType, targetParsedName, targetNode }) => {
+    if (targetNode.data.nodeType !== "tuple") return undefined;
+    if (targetParsedName === "value")
+      return sourceType === "list_number" || sourceType === "number";
+    if (targetParsedName.startsWith("row_")) return sourceType === "number";
+    return false;
+  },
+];
+
+// ---------------------------------------------------------------------------
+// Subgraph / boundary helpers
+// ---------------------------------------------------------------------------
+
+function checkSubgraphSource(
+  sourceNode: Node<FlowNodeData>,
+  sourceParsedName: string,
+  targetNode: Node<FlowNodeData>,
+  targetParsedName: string
+): boolean | undefined {
+  if (sourceNode.data.nodeType !== "subgraph") return undefined;
+  const port = sourceNode.data.subgraphOutputs?.find(
+    p => p.name === sourceParsedName
+  );
+  if (!port) return undefined;
+  if (port.paramType && targetNode.data.nodeType === "pipeline") {
+    const targetParam = targetNode.data.parameterInputs?.find(
+      p => p.name === targetParsedName
+    );
+    return targetParam ? port.paramType === targetParam.type : false;
+  }
+  return true;
+}
+
+function checkSubgraphTarget(
+  sourceNode: Node<FlowNodeData>,
+  targetNode: Node<FlowNodeData>,
+  targetParsedName: string,
+  nodes: Node<FlowNodeData>[]
+): boolean | undefined {
+  if (targetNode.data.nodeType !== "subgraph") return undefined;
+  const port = targetNode.data.subgraphInputs?.find(
+    p => p.name === targetParsedName
+  );
+  if (!port) return undefined;
+  const srcType = resolveSourceType(sourceNode, nodes, []);
+  if (!srcType) return true;
+  if (!port.paramType) return true;
+  return srcType === port.paramType;
+}
+
+function checkSubgraphInputSource(
+  sourceNode: Node<FlowNodeData>,
+  sourceParsedName: string,
+  targetNode: Node<FlowNodeData>,
+  targetParsedName: string
+): boolean | undefined {
+  if (sourceNode.data.nodeType !== "subgraph_input") return undefined;
+  const port = sourceNode.data.subgraphInputs?.find(
+    p => p.name === sourceParsedName
+  );
+  if (!port) return undefined;
+  if (port.paramType && targetNode.data.nodeType === "pipeline") {
+    const targetParam = targetNode.data.parameterInputs?.find(
+      p => p.name === targetParsedName
+    );
+    return targetParam ? port.paramType === targetParam.type : false;
+  }
+  return true;
+}
+
+function checkSubgraphOutputTarget(
+  sourceNode: Node<FlowNodeData>,
+  targetNode: Node<FlowNodeData>,
+  targetParsedName: string,
+  nodes: Node<FlowNodeData>[]
+): boolean | undefined {
+  if (targetNode.data.nodeType !== "subgraph_output") return undefined;
+  const port = targetNode.data.subgraphOutputs?.find(
+    p => p.name === targetParsedName
+  );
+  if (!port) return undefined;
+  const srcType = resolveSourceType(sourceNode, nodes, []);
+  if (!srcType) return true;
+  if (!port.paramType) return true;
+  return srcType === port.paramType;
+}
+
+// ---------------------------------------------------------------------------
+// Main exported validator
+// ---------------------------------------------------------------------------
+
+/**
+ * Returns `true` when the proposed connection should be allowed.
+ * Pure function – no React dependencies.
+ */
+export function validateConnection(
+  edgeOrConnection: Edge | Connection,
+  nodes: Node<FlowNodeData>[]
+): boolean {
+  let connection: Connection;
+  if ("source" in edgeOrConnection && "target" in edgeOrConnection) {
+    connection = edgeOrConnection as Connection;
+  } else {
+    const edge = edgeOrConnection as Edge;
+    connection = {
+      source: edge.source,
+      target: edge.target,
+      sourceHandle: edge.sourceHandle ?? null,
+      targetHandle: edge.targetHandle ?? null,
+    };
+  }
+
+  if (
+    !connection.source ||
+    !connection.target ||
+    !connection.sourceHandle ||
+    !connection.targetHandle
+  ) {
+    return true;
+  }
+
+  const sourceParsed = parseHandleId(connection.sourceHandle);
+  const targetParsed = parseHandleId(connection.targetHandle);
+
+  if (!sourceParsed || !targetParsed) return true;
+
+  // Boundary "add" ports always valid
+  if (
+    (connection.source === BOUNDARY_INPUT_ID &&
+      sourceParsed.name === "__add__") ||
+    (connection.target === BOUNDARY_OUTPUT_ID &&
+      targetParsed.name === "__add__")
+  ) {
+    return true;
+  }
+
+  // Stream ↔ stream always ok
+  if (sourceParsed.kind === "stream" && targetParsed.kind === "stream")
+    return true;
+
+  // Param ↔ param
+  if (sourceParsed.kind === "param" && targetParsed.kind === "param") {
+    const sourceNode = nodes.find(n => n.id === connection.source);
+    const targetNode = nodes.find(n => n.id === connection.target);
+    if (!sourceNode || !targetNode) return false;
+
+    // Primitive can connect to anything
+    if (sourceNode.data.nodeType === "primitive") return true;
+
+    // Reroute source
+    if (sourceNode.data.nodeType === "reroute") {
+      if (!sourceNode.data.valueType) return true;
+      if (targetNode.data.nodeType === "reroute") {
+        return (
+          !targetNode.data.valueType ||
+          targetNode.data.valueType === sourceNode.data.valueType
+        );
+      }
+    }
+
+    // Reroute target
+    if (targetNode.data.nodeType === "reroute") {
+      if (!targetNode.data.valueType) return true;
+      const srcType = resolveSourceType(sourceNode, nodes, []);
+      if (!srcType) return true;
+      return srcType === targetNode.data.valueType;
+    }
+
+    const sourceType = resolveSourceType(sourceNode, nodes, []);
+    if (!sourceType) return false;
+
+    // Run table-driven target rules
+    for (const rule of TARGET_RULES) {
+      const result = rule({
+        sourceType,
+        targetParsedName: targetParsed.name,
+        targetNode,
+      });
+      if (result !== undefined) return result;
+    }
+
+    // Subgraph / boundary node checks
+    const sgSrc = checkSubgraphSource(
+      sourceNode,
+      sourceParsed.name,
+      targetNode,
+      targetParsed.name
+    );
+    if (sgSrc !== undefined) return sgSrc;
+
+    const sgTgt = checkSubgraphTarget(
+      sourceNode,
+      targetNode,
+      targetParsed.name,
+      nodes
+    );
+    if (sgTgt !== undefined) return sgTgt;
+
+    const sgInSrc = checkSubgraphInputSource(
+      sourceNode,
+      sourceParsed.name,
+      targetNode,
+      targetParsed.name
+    );
+    if (sgInSrc !== undefined) return sgInSrc;
+
+    const sgOutTgt = checkSubgraphOutputTarget(
+      sourceNode,
+      targetNode,
+      targetParsed.name,
+      nodes
+    );
+    if (sgOutTgt !== undefined) return sgOutTgt;
+
+    // Default: look up the target's parameterInputs
+    const targetParam = targetNode.data.parameterInputs?.find(
+      p => p.name === targetParsed.name
+    );
+    if (!targetParam) return false;
+
+    if (targetParam.type === "list_number" && sourceType === "number")
+      return true;
+    if (targetParam.type === "list_number" && sourceType === "list_number")
+      return true;
+
+    return sourceType === targetParam.type;
+  }
+
+  return false;
+}
